test: refactoring of tests to comply with changes
