+++ service/src/test/java/org/jboss/sbomer/service/test/integ/feature/s3/S3FeatureIT.java
@@ -38,7 +38,7 @@ import org.jboss.sbomer.service.feature.sbom.k8s.model.GenerationRequest;
 import org.jboss.sbomer.service.feature.sbom.k8s.model.SbomGenerationStatus;
 import org.jboss.sbomer.service.feature.sbom.model.SbomGenerationRequest;
 import org.jboss.sbomer.service.test.integ.feature.s3.S3FeatureIT.S3ClientConfig;
-import org.jboss.sbomer.service.test.integ.feature.sbom.messaging.AmqpTestResourceLifecycleManager;
+import org.jboss.sbomer.service.test.utils.umb.TestUmbProfile;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
@@ -46,9 +46,7 @@ import org.junit.jupiter.api.io.TempDir;
 import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.quarkus.panache.mock.PanacheMock;
 import io.quarkus.test.InjectMock;
-import io.quarkus.test.common.WithTestResource;
 import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.QuarkusTestProfile;
 import io.quarkus.test.junit.TestProfile;
 import io.quarkus.test.junit.mockito.InjectSpy;
 import io.restassured.RestAssured;
@@ -58,7 +56,6 @@ import io.restassured.http.ContentType;
 
 @QuarkusTest
 @TestProfile(S3ClientConfig.class)
-@WithTestResource(AmqpTestResourceLifecycleManager.class)
 class S3FeatureIT {
 
     @InjectSpy
@@ -73,7 +70,7 @@ class S3FeatureIT {
     @InjectMock
     S3ClientFacade clientFacade;
 
-    public static class S3ClientConfig implements QuarkusTestProfile {
+    public static class S3ClientConfig extends TestUmbProfile {
         @Override
         public Map<String, String> getConfigOverrides() {
             return Map.of(
@@ -84,7 +81,9 @@ class S3FeatureIT {
                     "aws.access.key.id",
                     "access-key",
                     "aws.secret.access.key",
-                    "secret-access-key");
+                    "secret-access-key",
+                    "sbomer.features.umb.enabled",
+                    "true");
         }
     }
